Particles no longer smear over screen when crossing edge
The problem was that the particle system on the Ship prefab
emit particles when it detects movement. Unfortunately screen
wrapping confuse the particle system into emitting lots of
particles which smeared across the entire screen.

My solution was to put a script on the particle system that
listens for screen wraps and disables emission for one frame.