Simple Score UI
I created a simple, animated score UI. Using it should be as
simple as performing the following steps:

- Drop Score prefab into scene
- Optionally provide different audio clip
- Call Score.Earn(100); to express player earned 100 points.

This will fire off visual and audio feedback from ScoreUI.cs

Current score system:

 - 25 points for small asteroid
 - 100 points for large asteroid
 - 100*N bonus points for level cleared

Player scores whenever an asteroid is shot and when the level is
cleared the player is awarded a bonus. I placed some constants
in AsteroidBehaviour.cs which can be moved somewhere else. I also
placed a constant in GameManager.cs because it's the class that
currently is concerned with the details of clearing a level.

In the case Player dies, score is tallied up as a final step to
present the player with the result of their carnage. For now,
all it does is just flash the score feedback system exactly the
same way as when you get score shooting asteroids.